requirements.txt
Flask
pyembroidery
Pillow